source/requirements.txt
flask==3.0.0
prometheus-client==0.21.1
psycopg2-binary==2.9.9
gunicorn==23.0.0
